+++ lib/websockex.ex
@@ -72,7 +72,10 @@ defmodule WebSockex do
   @typedoc """
   The error returned when a connection fails to be established.
   """
-  @type connection_error :: %WebSockex.RequestError{} | %WebSockex.ConnError{}
+  @type close_error :: %WebSockex.RequestError{}
+                       | %WebSockex.ConnError{}
+                       | %WebSockex.InvalidFrameError{}
+                       | %WebSockex.FrameEncodeError{}
 
   @typedoc """
   A map that contains information about the failure to connect.
@@ -80,7 +83,7 @@ defmodule WebSockex do
   This map contains the error, attempt number, and the `t:WebSockex.Conn.t/0`
   that was used to attempt the connection.
   """
-  @type connection_status_map :: %{reason: close_reason | connection_error,
+  @type connection_status_map :: %{reason: close_reason | close_error,
                                    attempt_number: integer,
                                    conn: WebSockex.Conn.t}
 
@@ -308,7 +311,8 @@ defmodule WebSockex do
   @doc """
   Queue a frame to be sent asynchronously.
   """
-  @spec send_frame(pid, frame) :: :ok | {:error, %WebSockex.FrameEncodeError{}}
+  @spec send_frame(pid, frame) :: :ok | {:error, %WebSockex.FrameEncodeError{}
+                                                 | %WebSockex.InvalidFrameError{}}
   def send_frame(pid, frame) do
     with {:ok, binary_frame} <- WebSockex.Frame.encode_frame(frame) do
       send(pid, {:"$websockex_send", binary_frame})
@@ -569,6 +573,9 @@ defmodule WebSockex do
   defp handle_close({:local, _, _} = reason, parent, debug, state) do
     handle_local_close(reason, parent, debug, state)
   end
+  defp handle_close({:error, _} = reason, parent, debug, state) do
+    handle_error_close(reason, parent, debug, state)
+  end
 
   defp common_handle({function, msg}, parent, debug, state) do
     result = try_callback(state.module, function, [msg, state.module_state])
@@ -582,14 +589,15 @@ defmodule WebSockex do
               do: :ok = WebSockex.Conn.socket_send(state.conn, binary_frame)
         case res do
           :ok -> websocket_loop(parent, debug, %{state | module_state: new_state})
-          {:error, error} -> raise error
+          {:error, error} ->
+            handle_close({:error, error}, parent, debug, %{state | module_state: new_state})
         end
       {:close, new_state} ->
         handle_close({:local, :normal}, parent, debug, %{state | module_state: new_state})
       {:close, {close_code, message}, new_state} ->
         handle_close({:local, close_code, message}, parent, debug, %{state | module_state: new_state})
       {:"$EXIT", reason} ->
-        terminate(reason, parent, debug, state)
+        handle_error_close(reason, parent, debug, state)
       badreply ->
         error = %WebSockex.BadResponseError{module: state.module, function: function,
           args: [msg, state.module_state], response: badreply}
@@ -616,14 +624,30 @@ defmodule WebSockex do
     end
   end
 
+  defp handle_error_close(reason, parent, debug, state) do
+    send_close_frame(:error, state.conn)
+
+    Process.send_after(self(), :"$websockex_close_timeout", 5000)
+    close_loop(reason, parent, debug, state)
+  end
+
+  def handle_terminate_close(reason, parent, debug, state) do
+    send_close_frame(:error,  state.conn)
+
+    # I'm not supposed to do this, but I'm going to go ahead and close the
+    # socket here. If people complain I'll come up with something else.
+    new_conn = WebSockex.Conn.close_socket(state.conn)
+    terminate(reason, parent, debug, %{state | conn: new_conn})
+  end
+
   # Frame Sending
 
   defp handle_send(binary_frame, parent, debug, %{conn: conn} = state) do
     case WebSockex.Conn.socket_send(conn, binary_frame) do
       :ok ->
-      websocket_loop(parent, debug, state)
+        websocket_loop(parent, debug, state)
       {:error, error} ->
-      terminate(error, parent, debug, state)
+        handle_close(error, parent, debug, state)
     end
   end
 
@@ -637,6 +661,9 @@ defmodule WebSockex do
   end
   defp build_close_frame({_, code, msg}) do
     WebSockex.Frame.encode_frame({:close, code, msg})
+  end
+  defp build_close_frame(:error) do
+    WebSockex.Frame.encode_frame({:close, 1011, ""})
   end
 
   # Connection Handling
@@ -660,6 +687,8 @@ defmodule WebSockex do
 
   defp on_disconnect(reason, parent, debug, state, attempt \\ 1) do
     case handle_disconnect(reason, state, attempt) do
+      {:ok, new_module_state} when is_tuple(reason) and elem(reason, 0) == :error ->
+        terminate(elem(reason, 1), parent, debug, %{state | module_state: new_module_state})
       {:ok, new_module_state} ->
         terminate(reason, parent, debug, %{state | module_state: new_module_state})
       {:reconnect, new_conn, new_module_state} ->
@@ -734,6 +763,10 @@ defmodule WebSockex do
     end
   end
 
+  defp terminate(reason, parent, debug, %{conn: %{socket: socket}} = state)
+  when not is_nil(socket) do
+    handle_terminate_close(reason, parent, debug, state)
+  end
   defp terminate(reason, _parent, _debug, %{module: mod, module_state: mod_state}) do
     mod.terminate(reason, mod_state)
     case reason do